Remove remaining global config settings in favor of site-specific settings.
A positive side effect of this change is that the api_request_delay now
operates on a per-domain basis. Thus there are different rate limits for
different sites that you are connecting to. Note that the key for the last_call
is the domain name, thus even though the Reddit or config object may be
different, if the domain is the same, then the rate limit will apply.