Merge docstring improvments from Damgaard master.
In addition to excellent docstring improvements, Damgaard's changes brought
forth a rename of the `compose` function to `send`. The compose function will
produce a deprecation warning until it is finally removed. The estimated
removal time will be the end of 2012.

Additions to the merge include lint corrections (trailing whitespaces and line
lenghth issues) and tests to assert a deprecation warning is generated when
using the compose function.