Reworked login to work with praw.ini.
The Docstring said that a user could modify praw.ini to allow for
automatic login. How to do that was both unclear and had bug potential,
the login function would only function if user/pswd was either missing
or included a username/password. If it was empty, then it would accept
the empty string as the username/password rather than query user.