Make consistent the use of params and data.
Previously `url_data` was used for what is now called `params`. This is data
that is passed as part of the URL in an HTTP request.

More confusingly, previously `params` was used to refer to data passed in part
of the body of an HTTP request (POSTs exclusively). These are now consistently
named `data`.