Add and use alias_function factory for simplification of function aliases.
In effort to both minimize the amount of code, and reduce maintaining similar
docstrings, the RedditContentObjects now have an easy way to alias methods from
their BaseReddit extensions. Conveniently this provides them with the identical
docsting (when inspecting) thus we just need to maintain a single
docstring. The downside is that the BaseReddit extension version of each
function requires an additional argument (the RedditContentObject), however,
that's pretty easy to correct for in the documentation (we'll have to do that).

As part of this change, I removed the ability to (un)subscribe to subreddits by
id. That functionality was previously provided only through the `_subscribe`
method thus it wasn't part of PRAW's public "interface".