Add and pass mysubreddits scope test.
Doing this required moving my_contributions, my_moderation, and my_reddits into
the Reddit namespace. A MySubredditMixin class was created specifically for
these items. Furthermore, for consistency, each of the functions is now
prefixed with get_.

In testing this it came to light that the _use_oauth attribute doesn't work
with generators as the code in the generator is only evaluated when items are
needed from the generator. Thus I created a oauth_generator decorator and
updated the get_content generator to work with oauth.