Remove all_comments property. Use replace_more_comments.
The all_comments property was somewhat confusing. Calling the now public
`replace_more_comments` function is now an explicit action that allows for
fine-tuning via its `limit` and `threshold` paramters.

Before to replace instances of more comments the user needed only to do:

    s.all_comments

now the user needs to explicitly replace the more objects

    skipped = s.replace_more_comments()
    s.comments

As part of this change the max_more_comments config value has been removed.

Finally, bumped the version to 2.0rc0 as this should be the last breaking
change.