Make Subreddit.get_comments part of _methods
This allows it to be used by the Reddit object, similair to get_flair
or get_banned. This change means helpers._get_section is now only used
to generate non-subreddit listings as per it's docstring.

Also make Reddit.get_all_comments call Reddit.get_comments with 'all'
as the subreddit.