Final cleanup of test PR
Removed all instances of six.next. PRAW works for python 2.6+
and the builtin function next was introduced in python 2.6, so
we can just use the builtin.

Removed docstring in praw/tests/helper. It was the left-over
docstring from the original big test file. There is currently
no module docstrings for any of the testing files. Their title
should be sufficient.

Fix misspellings of logged in 3 functions in access_control_test