Move request handling (+rate-limit and caching) into handlers.py.
Very soon, `Reddit` instances will be able to be created with separate
handlers. As such, this commit moves the existing request handling code into a
`DefaultHandler` under `handlers.py`.

At present the only solid part of the handler interface is that a handler
provides three functions:

* clear_cache
* evict
* request

Also, bump the `requests` package number, and fix a number of lint warnings.