Add handler hierarchy. Make RateLimitHandler thread safe.
The RateLimit handler is now thread safe. Once the Default handler with caching
support is also made thread safe, then PRAW will be thread safe with respect to
a one-to-one mapping between threads and `Reddit` instances.

The thread safety in the RateLimitHandler.rate_limit decorator gives the
multiprocess worker thread safety.

As part of this change, the `clear_cache` has been removed. The cache was
previously emptied on authentication change, however, that was unecessary.

All of these changes are in internal PRAW code, and thus require no changelog
update. However, once finalized a changelog entry should be made briefly
describing the addition of the `handler` argument to `Reddit`.