Extend MultiprocessHandler error handling.
Actually raise exceptions "returned" from the multiprocess server.

The MultiprocessHandler will retry indefinitely (up to 64 second intervals) to
connect to the multiprocess server. This includes graceful handling of dropped
connections.

Permit a limited number of retries (3) for single request that results in a
successfully connection to the multiprocess server followed by a dropped
connection. Such behavior would indicate that the request is crashing the
multiprocess server thread responsible for handling the request. In such a case
a PRAW ClientException is raised.