Handle aliased calls where subreddit isn't 1st arg
The inability before to handle this caused a bug in
Subreddit.search. The search method in UR, which this method
aliasses, takes the query as the first argument and the subreddit
as the second. alias would then swap these variables causing
the method call to (usually) give a 404 error as it would try to
find a subreddit with the name of the query.

test_search in praw/tests/__init__.py now properly passes.