Rename _get_section->_get_redditor_listing.
This method is only used to return redditor listings. If
reddit adds a non-subreddit listing, then this function
might also be usable for that. However, there is a no way
to know when or if reddit will implement such a feature
nor whether this function might certainly work with it.

The benefits of this name therefore doesn't outweigh the
readability disadvantages by having a function with a name
that make it sound far more general than it actually is.