Expect test_get_friends to raise RedirectException
get_friends is a bit tricky to test, as it awaits a fix upstream.
See issue #175 Until that happens, it will remain broken. However,
having a test where the expected behaviour is to fail is wrong.

This commit mean that in the future, the problem will be solved and
this test will now fail in this version of PRAW. Therefore,
documentation has been added to show that this is actually not a
problem to any users running the test suite for this version in the
future.