Run tests using tox on all supported Python interpreters.
If a user does not have the required interpreter installed, it will be
skipped.

Set REDDIT_SITE envvar when running tox

The Travis config specifies this environment variable; I'm not sure
exactly what it does, but we'll set it, too, when running tests with
Tox.

The full test suite does not actually pass. The Travis configuration
specifies only a subset of test suites that should be run, so we'll run
the same tests when running Tox.