Update many of the subreddit image related tests.
Two image upload tests are included which are not cached via betamax. The
primary reason for this is that because of randomness in subsequent requests
pertaining to multipart encoding, the matcher does not detect the matching
requests. Thus the custom matcher must be expanded.

Once that works, the other methods should be uncommented, and then cached via
betamax.