Hide debug ouput on travis for submissions_between
The test on travis for submissions_between is putting out debug outputs at a level of 3 and less. While great for live changes and seeing what goes wrong when making them; on travis it just kinda spams the coverage log. So; on travis; this commit makes the debug ouputs hidden.