Add mixins/listing.py and models/listinggenerator.py.
The mixins.Listing class now provides the top, new, hot, etc, functions for
objects that represent such listings. This change obviates `_get_sorter`, so
that was removed from internal.py.

PRAWModel is the new superclass for all PRAW-related models. This additional
superclass is needed as not all PRAWModels map to a data-type returned by
reddit (e.g., the frontpage, which is not a Subreddit).

The functionality that was previously provided by `get_content` is now provided
by the ListingGenerator class.