Only perform the update check once per session
Currently, the update checker will run each time a `Reddit` instance is created. This is because accessing `update_checked` with `self.` treats it as an instance variable rather than a static class variable. Changing the call to `[class name].update_checked` will treat it as a static class variable.

I haven't tested this change (so make sure to test it please), but a similar patch worked in my code base.