Add pyopenssl as a conditional dependency
While pyopenssl is not actually required, requests injects it into urllib3 if it exists on the system (as is the default on some Linux distros). However, some versions < 0.15 have a regression in which str/unicode is not accepted for certificate verification.

So, the simplest solution is to require pyopenssl conditionally.