Add docs/redirect_log.txt
Readthedocs supports manually entering page redirects. Given the tedious
process to create the list, it makes sense to keep a back-up list in case those
redirects are somehow lost on readthedocs.

The redirects are necessary to provide a user-friendly interface for users when
switching back and forth between versions of the documentation that span the
PRAW3/PRAW4 boundary because all pages save for `/` were renamed.