Added example to CommentModeration
This is my first praw edit so let me know if this is incorrect, or if I've not followed your code style guidelines (I couldn't find any). I'm u/theonefoster on reddit. 

I also don't know of a way to test this commit to make sure it formats correctly when published so this code is untested.