Fix #1271.
NoReturn is not included in some versions of Python 3.5 and 3.6.
Further, the annotation was misused. NoReturn indicates that a function
_will not_ return, either because it unconditionally throws an exception
or because it goes into an infinite loop or because it terminates
program execution. The functions annotated with NoReturn here do not
have that behavior, though each of them *may* throw an exception. At
best, these functions should have been annotated with a union of None
and NoReturn, but since the NoReturn name is problematic, it is best to
omit it entirely.