ListingGenerator is an Iterator; Generator -> Iterator
In many places, functions claim to return a Generator,
which (per my understanding of the documentation) is a
stronger notion than just an Iterator.

For ListingGenerator to be a proper Generator, it would
need to implement some abstract methods. In contrast, it
already satisfies the definition of an Iterator.

Since our generators never accept a send type, and never
return something, they are equivalent to Iterator[type].

https://docs.python.org/3/library/typing.html#typing.Generator