Add LiveThreadStream class to stream live thread updates.
Streaming live thread updates seems like a pretty obvious feature
given the "streaming" nature of a live thread.

The interface defined here is to use a stream via:

for update in reddit.live('something').stream.updates():
    print(update)

I considered an interface along the lines of

for update in reddit.live('something').stream():
    print(update)

but decided on the other way because the other way is more consistent
with the existing interface to stream subreddits.