Fix pre-push script to work on Windows
Since Windows and all Unix-based OSes have different conventions on process initialization, check_call needs to always be a list, so that Python can accomondate each operating system. The changed line translates as "execute a file named `black ..exe`", not "execute a file named `black.exe` with a parameter of `.`, as it does on Unix. The fix just converts the list of arguments to an actual list of arguments.