Add integration tests for info(sr_names)
In accordance with some contributing guidelines, this commit adds the two integration tests needed.

The test uses one subreddit name that is normally formatted (/r/redditdev) and three others, which while are still valid subreddits, are different from the norm in some way (/r/nl's name is too short, /r/t:1337 is a timereddit which has a colon in its subreddit name, and /r/reddit.com which has a period in its name.

Each subreddit exists and is publicly accessible, which should make tests that use these subreddit succeed for the forseeable future. If the visibility of any of the subreddits do change, the test may have to be changed to reflect this.