Allow Subreddit objects to be passed into Reddit.info.
This change allows Subreddit objects to be passed into the Reddit.info() function when using subreddits. I've also changed the name `sr_names` to `subreddits`.

In addition, the "mutually exclusive paramters" error has been replaced with one single check that checks if the amount of parameters passed to the function are correct.

The tests (both the integration and unit tests) have been updated to cover these cases.