Adds the "subreddit" parameter to Reddit.info(). (#1570)
This change adds the `subreddit` parameter as defined in the reddit API documentation (https://www.reddit.com/dev/api#GET_api_info). This change was implemented in the Reddit API earlier today, which makes it easier for developers to get many subreddits at once by their name, which can significantly reduce the amount of requests made by an API client.

This change adds the parameter to info, updates relevant checks, and updates the test that test for IDs provided. Previously, if developers wanted to use this new parameter they would have to perform the request manually.

In accordance with some contributing guidelines, this commit adds the two integration tests needed.

The test uses one subreddit name that is normally formatted (/r/redditdev) and three others, which while are still valid subreddits, are different from the norm in some way (/r/nl's name is too short, /r/t:1337 is a timereddit which has a colon in its subreddit name, and /r/reddit.com which has a period in its name.

Each subreddit exists and is publicly accessible, which should make tests that use these subreddit succeed for the forseeable future. If the visibility of any of the subreddits do change, the test may have to be changed to reflect this.