Connect to websocket before submitting media
The Reddit API has been updated since this feature was originally
written. Now, we can get a websocket URL from our media upload request.
This enables us to connect to the websocket before even making our
submission. This in turn makes it much harder to miss a message on the
websocket, especially in slow network conditions, because we're no
longer racing to connect to the websocket immediately after submitting
our post.

Thanks to Shaun Pedicini (@scpedicini on GitHub) for pointing this out
to me in a private email.