Update Config to use importlib for defaults
Makes praw "zip-safe" by using importlib to read the default praw.ini
instead of trying to build a path to the ini file on disk.

Prior to this patch, attempting to use `praw` in a zipapp will exit with
an exception because the `praw.ini` file does not actually exist in the
system filesystem.